Fixing all models to match up with our current data model

+++ src/main/java/team/awesome/pet/dao/ShelterDao.java
@@ -1,5 +1,6 @@
 package team.awesome.pet.dao;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 import team.awesome.pet.model.Shelter;
@@ -13,5 +14,5 @@ public interface ShelterDao {
 
   void insertShelterToDb(Shelter shelter);
 
-  void removeShelterById(int id);
+  void removeShelterById(BigInteger id);
 }
